Fixed the null images upload in Panel

+++ app/Http/Controllers/Panel.php
@@ -180,18 +180,22 @@ class Panel extends Controller
             $image = base_path() . '/public/temp/' . $_POST['cover_img'];
             $s3 = \Storage::disk('s3');
             $filePath = '/images/' . $_POST['cover_img'];
-            if($s3->put($filePath, file_get_contents($image), 'public')){
-                $user->coverImageUrl = $_POST['cover_img'];
-                unlink($image);
+            if (file_exists($image) ) {
+              if($s3->put($filePath, file_get_contents($image), 'public')){
+                  $user->coverImageUrl = $_POST['cover_img'];
+                  unlink($image);
+              }
             }
         }
         if(isset($_POST['profile_img']) && !empty($_POST['profile_img'])){
             $image = base_path() . '/public/temp/' . $_POST['profile_img'];
             $s3 = \Storage::disk('s3');
             $filePath = '/images/' . $_POST['profile_img'];
-            if($s3->put($filePath, file_get_contents($image), 'public')){
-                $user->companyLogoUrl = $_POST['profile_img'];
-                unlink($image);
+            if (file_exists($image)) {
+              if($s3->put($filePath, file_get_contents($image), 'public')){
+                  $user->companyLogoUrl = $_POST['profile_img'];
+                  unlink($image);
+              }
             }
         }
 
@@ -363,18 +367,25 @@ class Panel extends Controller
 
         // TODO: handle for images already in S3
         if (isset($_POST['images']) && count($_POST['images']) > 0 ) {
+            foreach($_POST['images'] as $i => $val) {
+              if ( !isset($val) || empty($val) || strtolower($val) === 'null') {
+                  unset($_POST['images'][$i]);
+              }
+            }
             $s3 = \Storage::disk('s3');
             $uploadedImage = array();
-            for($i =0 ;  $i < count($_POST['images']); $i++) {
-                if (!$s3->has('/images/'. $_POST['images'][$i])) {
-                    $image = base_path() . '/public/temp/' . $_POST['images'][$i];
-                    $filePath = '/images/' . $_POST['images'][$i];
-                    if($s3->put($filePath, file_get_contents($image), 'public')){
-                        $uploadedImage[] = $_POST['images'][$i];
-                        unlink($image);
+            foreach( $_POST['images'] as $i => $val) {
+                if (!$s3->has('/images/'. $val)) {
+                    $image = base_path() . '/public/temp/' . $val;
+                    $filePath = '/images/' . $val;
+                    if ( file_exists($image)) {
+                      if($s3->put($filePath, file_get_contents($image), 'public')){
+                          $uploadedImage[] = $val;
+                          unlink($image);
+                      }
                     }
                 } else {
-                    $uploadedImage[] = $_POST['images'][$i];
+                    $uploadedImage[] = $val;
                 }
             }
 
@@ -468,19 +479,23 @@ class Panel extends Controller
             $image = base_path() . '/public/temp/' . $_POST['cover_img'];
             $s3 = \Storage::disk('s3');
             $filePath = '/images/' . $_POST['cover_img'];
-            if($s3->put($filePath, file_get_contents($image), 'public')){
-                $user->coverImageUrl = $_POST['cover_img'];
-                unlink($image);
+            if (file_exists($image)) {
+              if($s3->put($filePath, file_get_contents($image), 'public')){
+                  $user->coverImageUrl = $_POST['cover_img'];
+                  unlink($image);
+              }
             }
         }
         if(isset($_POST['profile_img']) && !empty($_POST['profile_img'])){
             $image = base_path() . '/public/temp/' . $_POST['profile_img'];
             $s3 = \Storage::disk('s3');
             $filePath = '/images/' . $_POST['profile_img'];
-            if($s3->put($filePath, file_get_contents($image), 'public')){
-                $user->companyLogoUrl = $_POST['profile_img'];
-                unlink($image);
-            }
+            if (file_exists($image)) {
+              if($s3->put($filePath, file_get_contents($image), 'public')){
+                  $user->companyLogoUrl = $_POST['profile_img'];
+                  unlink($image);
+              }
+            } 
         }
 
 
@@ -633,7 +648,7 @@ class Panel extends Controller
       ->join('currencies', 'listings.currencyId', '=', 'currencies.id')
       ->join('users', 'listings.userId', '=', 'users.id')
       ->orderby('listings.created_at', 'desc')
-      ->select('listings.*', 'currencies.code', 'users.*')
+      ->select('listings.*', 'currencies.code', 'users.firstName', 'users.lastName', 'users.username', 'users.companyName', 'users.fullName')
       ->where('status', '<>', 'EXPIRED')
       ->paginate(10);
